Email file grabber
This should grab the most recent excel file of the specified eMARS stuff. Then download this into a separate folder and copy this data into a main excel worksheet. The hard part is our excel sheet has some formulas that references previous sheets, which are all named after a date of the month. So, had to introduce some datetime logic so that it would input the correct date into the formula after updating. Should be good to be ran daily and autonomically.