Email grabbing code
This is similar to the daily_balances.py file except that it filters the downloaded sheet from email by the POOL column and only takes the "L". Then it also inserts it into a different cash balance folder and excel spreadsheet. It does the same J column manipulation with the dates but only one other cell needs the equation changed for the date.